% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-package.R
\docType{package}
\name{flow-package}
\alias{flow-package}
\title{Data analysis workflow and alternate pipeline operator}
\description{
Data (work)flow analysis using proto objects and pipe operator that
integrates non-standard evaluation and the lazyeval mechanism.
}
\section{Important functions}{

\itemize{
\item \code{\link[=flow]{flow()}} constructs a \code{Flow} object.
\item \code{\link[=as_quosure]{as_quosure()}}, and unary \code{+} and \code{-} operators combined with \strong{formula}
objects provide an alternate way to manipulate \strong{quosure}s.
\item \code{\link[=quos_underscore]{quos_underscore()}} automatically convert arguments whose name ends with
\code{_} into \strong{quosure}s, and this mechanism is used by our pipe operators.
\item \code{\link[=debug_flow]{debug_flow()}} provides a convenient way to debug problematic pipelines
build with our own pipe operators \code{\link{\%>.\%}} and
\code{\link{\%>+\%}}.
}
}

