% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_flow.R
\name{graph_flow}
\alias{graph_flow}
\title{Create a graph with Flow objects hierarchy}
\usage{
graph_flow(env = .GlobalEnv, child_to_parent = TRUE, plotit = TRUE, ...)
}
\arguments{
\item{env}{The environment to look for \strong{Flow} objects. By default, it is
\code{.GlobalEnv}, and you should not change it, since all \strong{Flow} objects are
derived from it by construction.}

\item{child_to_parent}{Do the arrows go from child to parent (by default), or
in the other direction?}

\item{plotit}{Do we plot the graph (by default)?}

\item{...}{Further parameters passed to \code{\link[=plot.igraph]{plot.igraph()}}.}
}
\value{
An \strong{igraph} object (returned invisibly if \code{plotit = TRUE}.
}
\description{
A graph showing all \strong{Flow} objects heritage is calculated, and displayed.
}
\examples{
a <- flow()
b <- a$flow()
c <- b$flow()
d <- a$flow()
# Use of custom names
e <- flow(.name = "parent")
f <- e$flow(.name = "child")
graph_flow()

# Arrows pointing from childs to parents, and do not plot it
g <- graph_flow(child_to_parent = FALSE, plotit = FALSE)
g
plot(g)
}
\seealso{
\link{flow}
}
\concept{
display objects hierarchy
}
\keyword{utilities}
