% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quos_underscore.R
\name{quos_underscore}
\alias{quos_underscore}
\title{Convert arguments whose names end with _ into quosures automatically}
\usage{
quos_underscore(...)
}
\arguments{
\item{...}{The named arguments provided to be either converted into quosures
or evaluated.}
}
\value{
An object of class \strong{quosures} is returned. It can be used directly
in tidyeval-aware contexts.
}
\description{
The expressions provided for all arguments whose names end with \verb{_} are
automatically converted into \strong{quosure}s, and also assigned to a name
without the training \verb{_}. The other arguments are evaluated in an usual way.
}
\examples{
foo <- function(...)
  quos_underscore(...)
foo(x  = 1:10, # "Normal" argument
    y_ = 1:10, # Transformed into a quosure
    z_ = non_existing_name) # Expressions in quosures are not evaluated
}
\seealso{
\link{as_quosure}, \code{\link{\%>_\%}}
}
\concept{automatic quosures creation for non-standard evaluation}
\keyword{utilities}
