% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.R
\name{str.Flow}
\alias{str.Flow}
\title{Compactly display the content of a Flow object}
\usage{
\method{str}{Flow}(object, max.level = 1, nest.lev = 0,
  indent.str = paste(rep.int(" ", max(0, nest.lev + 1)), collapse = ".."),
  ...)
}
\arguments{
\item{object}{A \strong{Flow} object.}

\item{max.level}{The maximum nesting level to use for displaying nested
structures.}

\item{nest.lev}{Used internally for pretty printing nested objects (you
probably don't want to change default value).}

\item{indent.str}{Idem.}

\item{...}{Further arguments passed to \code{str()} methods of \strong{Flow} items.}
}
\description{
Print short informative strings about the \strong{Flow} object and all it
containts, plus possibly, inheritage information.
}
\examples{
# TODO...
}
\seealso{
\code{\link[=flow]{flow()}}
}
\concept{
compactly inform about an object
}
\keyword{utilities}
