% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quosure.R
\name{quos_underscore}
\alias{quos_underscore}
\title{Convert names ending with _ into quosures automatically}
\usage{
quos_underscore(...)
}
\arguments{
\item{...}{The named arguments provided to be either converted into quosures
or evaluated.}
}
\description{
All names that end with _ are automatically converted into quosures, and also
assigned to the name without training \code{_}. The other arguments are evaluated.
}
\examples{
# TODO...
}
\seealso{
\code{\link[=as_quosure]{as_quosure()}}, \code{\link{\%>_\%}}
}
\concept{
automatic quosures creation for non-standard evaluation
}
\keyword{utilities}
