% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.Flow.R
\name{str.Flow}
\alias{str.Flow}
\title{Compactly display the content of a Flow object}
\usage{
\method{str}{Flow}(
  object,
  max.level = 1L,
  nest.lev = 0L,
  indent.str = paste(rep.int(" ", max(0L, nest.lev + 1L)), collapse = ".."),
  ...
)
}
\arguments{
\item{object}{A \strong{Flow} object.}

\item{max.level}{The maximum nesting level to use for displaying nested
structures.}

\item{nest.lev}{Used internally for pretty printing nested objects (you
probably don't want to change default value).}

\item{indent.str}{Idem.}

\item{...}{Further arguments passed to \code{str()} methods of \strong{Flow} items.}
}
\description{
Print short informative strings about the \strong{Flow} object and all it
contains, plus possibly, inheritance information.
}
\examples{
# A Flow object
data(iris)
fl <- flow(iris, x = 1:10, var_ = Sepal.Length)
fl # Shows the .value contained into fl
str(fl) # Provides compact information about satellite data contained in fl
}
\seealso{
\link{flow}
}
\concept{compactly inform about an object}
\keyword{utilities}
